feat: sign in and register button added, also functions to buttons
Added Sign in and Register button and added to all buttons on site functions that they move to their dedicated site.
