+++ frontend/src/App.tsx
@@ -1,10 +1,19 @@
-import MainMenu from "./scenes/MainMenu";
+import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
+import MainMenu from "./scenes/mainMenuscreen";
+import GameScreen from "./scenes/gameplayScreen";
+import SignIn from "./scenes/signinForm";
+import RegisterForm from "./scenes/RegisterForm";
 
 function App() {
   return (
-    <div>
-      <MainMenu />
-    </div>
+    <Router>
+      <Routes>
+        <Route path="/" element={<MainMenu />} />
+        <Route path="/game" element={<GameScreen />} />
+        <Route path="/signin" element={<SignIn />} />
+        <Route path="/register" element={<RegisterForm />} />
+      </Routes>
+    </Router>
   );
 }
