+++ frontend/src/scenes/gameplayScreen.tsx
@@ -1,0 +1,8 @@
+function GameScreen() {
+  return (
+    <div className="flex flex-col items-center justify-center h-screen bg-gray-900 text-white gap-4">
+      <h1>This is where the gameplay opens</h1>
+    </div>
+  );
+}
+export default GameScreen;
