+++ frontend/src/scenes/registerForm.tsx
@@ -1,0 +1,8 @@
+function RegisterForm() {
+  return (
+    <div className="flex flex-col items-center justify-center h-screen bg-gray-900 text-white gap-4">
+      <h1>This is where the register form opens</h1>
+    </div>
+  );
+}
+export default RegisterForm;
