+++ frontend/src/scenes/signinForm.tsx
@@ -1,0 +1,8 @@
+function SignIn() {
+  return (
+    <div className="flex flex-col items-center justify-center h-screen bg-gray-900 text-white gap-4">
+      <h1>This is where the sign in form opens</h1>
+    </div>
+  );
+}
+export default SignIn;
